#128 refactor: timer common 훅으로 분리
