+++ src/page/signUp/info/component/InfoForm/InfoForm.tsx
@@ -10,7 +10,6 @@ import { EMAIL_EXPIRED_MESSAGE, EMAIL_REMAIN_TIME, PLACEHOLDER, SUPPORTING_TEXT 
 import { useSendMailMutation } from '@/page/signUp/info/hook/api/useSendMailMutation';
 import { useDateInput } from '@/page/signUp/info/hook/common/useDateInput';
 import { useSelect } from '@/page/signUp/info/hook/common/useSelect';
-import { useTimer } from '@/page/signUp/info/hook/common/useTimer';
 import { formatTime } from '@/page/signUp/info/util/formatTime';
 
 import React, { useContext, useState } from 'react';
@@ -24,6 +23,7 @@ import Input from '@/common/component/Input/Input';
 import Select from '@/common/component/Select/Select';
 import { useOutsideClick, useOverlay } from '@/common/hook';
 import { useInput } from '@/common/hook/useInput';
+import { useTimer } from '@/common/hook/useTimer';
 
 const InfoForm = () => {
   const { isOpen, close, toggle } = useOverlay();
@@ -49,7 +49,7 @@ const InfoForm = () => {
   const context = useContext(SignUpContext);
   const navigate = useNavigate();
   const [isVerified, setIsVerified] = useState(false);
-  const mutate = useSendMailMutation(email, handleSend);
+  const mutate = useSendMailMutation(email);
   const { mutate: verifyCode } = useVerifyCodeMutation(email, authCode);
 
   const handleMailSend = () => {
