+++ src/page/signUp/info/hook/api/useSendMailMutation.ts
@@ -5,14 +5,11 @@ import { isAxiosError } from 'axios';
 import { postEmail } from '@/shared/api/mail/checking';
 import { useToastStore } from '@/shared/store/toast';
 
-export const useSendMailMutation = (email: string, trigger: () => void) => {
+export const useSendMailMutation = (email: string) => {
   const { createToast } = useToastStore();
 
   const { mutate: sendMailMutation } = useMutation({
     mutationFn: () => postEmail(email),
-    onMutate: async () => {
-      trigger();
-    },
     onError: (error) => {
       if (isAxiosError<{ message: string }>(error)) {
         createToast(`${error.response?.data.message}`, 'error');
